Updated makefile and code cleanup
