+++ Makefile
@@ -1,18 +1,67 @@
-BINARY=dist/gserver
+.PHONY: build fmt lint run test vendor_clean vendor_get vendor_update vet
+
+# Prepend our _vendor directory to the system GOPATH
+# so that import path resolution will prioritize
+# our third party snapshots.
+GOPATH := ${PWD}/_vendor:${GOPATH}
+export GOPATH
+
+
+BINARY=./bin/gserver
 
 VERSION=1.2.0
 BUILD=`git rev-parse HEAD`
 
 LDFLAGS=-ldflags "-X main.Version=${VERSION} -X main.Build=${BUILD}"
 
-.DEFAULT_GOAL: $(BINARY)
+.DEFAULT_GOAL: build
 
-$(BINARY):
-	go build ${LDFLAGS} -o ${BINARY}
+build:
+	go build ${LDFLAGS} -o ${BINARY} ./src/*.go
+
+run: build
+	${BINARY}
+
+# http://golang.org/cmd/go/#hdr-Run_gofmt_on_package_sources
+fmt:
+	go fmt ./src/...
+
+# https://github.com/golang/lint
+# go get github.com/golang/lint/golint
+lint:
+	golint ./src
+
+test:
+	go test ./src/...
 
 install:
-	go install ${LDFLAGS} -o ${BINARY}
+	go install ${LDFLAGS} -o ${BINARY} ./src/*.go
 
 clean:
 	if [ -f ${BINARY} ] ; then rm ${BINARY} ; fi
 
+vendor_clean:
+	rm -dRf ./_vendor/src
+
+# We have to set GOPATH to just the _vendor
+# directory to ensure that `go get` doesn't
+# update packages in our primary GOPATH instead.
+# This will happen if you already have the package
+# installed in GOPATH since `go get` will use
+# that existing location as the destination.
+vendor_get: vendor_clean
+	GOPATH=${PWD}/_vendor go get -d -u -v \
+	github.com/gorilla/context \
+	github.com/gorilla/websocket \
+	github.com/gorilla/mux
+
+vendor_update: vendor_get
+	rm -rf `find ./_vendor/src -type d -name .git` \
+	&& rm -rf `find ./_vendor/src -type d -name .hg` \
+	&& rm -rf `find ./_vendor/src -type d -name .bzr` \
+	&& rm -rf `find ./_vendor/src -type d -name .svn`
+
+# http://godoc.org/code.google.com/p/go.tools/cmd/vet
+# go get code.google.com/p/go.tools/cmd/vet
+vet:
+	go vet ./src/...
