+++ build-osx.sh
@@ -3,4 +3,4 @@ export GOOS="darwin"
 export GOARCH="amd64"
 export VERSION=1.2.0
 export BUILD=`git rev-parse HEAD`
-go build -ldflags "-X main.Version=${VERSION} -X main.Build=${BUILD}" -o dist/gserver-osx gserver.go
+go build -ldflags "-X main.Version=${VERSION} -X main.Build=${BUILD}" -o ./bin/gserver-osx ./src/gserver.go
