+++ build-win.sh
@@ -4,4 +4,4 @@ export GOARCH="amd64"
 export VERSION=1.2.0
 export BUILD=`git rev-parse HEAD`
 
-go build -ldflags "-X main.Version=${VERSION} -X main.Build=${BUILD}" -o dist/gserver.exe gserver.go
+go build -ldflags "-X main.Version=${VERSION} -X main.Build=${BUILD}" -o ./bin/gserver.exe ./src/gserver.go
