+++ src/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="utf-8">
+        <title>Simple Server</title>
+    </head>
+    <body>
+        <h2>Simple Server works!</h2>
+    </body>
+</html>
