+++ src/websocket.html
@@ -1,0 +1,72 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <title>WebSockets</title>
+    <style type="text/css">
+        body {
+            font-family:Helvetica,Arial,Tahoma
+        }
+        input[type="text"] {
+            padding: 4px;
+            min-width: 250px;
+        }
+
+        button { padding: 5px; }
+
+        .form {
+            float: left;
+            width: 300px;
+            min-height: 230px;
+            border: 1px solid #000000;
+        }
+        .tracelog {
+            float: left;
+            width: 300px;
+            min-height: 230px;
+            border: 1px solid #000000;
+            margin-left: 10px;
+            padding: 5px;
+        }
+        .field {
+            display: block;
+            padding: 10px;
+        }
+    </style>
+</head>
+<body>
+    <h2>Echo WebSocket Test</h2>
+    <script type="text/javascript" src="data/websocket.js"></script>
+    <br/>
+
+    <div>
+        <div class="form">
+            <form id="">
+                <div class="field">
+                    <button id="btnConnect" type="button">Connect</button>
+                    <button id="btnDisconnect" type="button" disabled="disabled">Disconnect</button>
+                </div>
+
+                <div class="field">
+                    <label for="location">Location</label><br/>
+                    <input type="text" id="location" name="location" maxlength="50" value="ws://localhost:9000/echo" />
+                </div>
+
+                <div class="field">
+                    <label for="message">Message</label><br/>
+                    <input type="text" id="message" name="message" maxlength="50" />
+                </div>
+
+                <div class="field">
+                    <button type="button" id="btnSend" disabled="disabled">Send</button>
+                </div>
+            </form>
+        </div>
+
+        <div class="tracelog">
+            <div id="output"><div>
+        </div>
+    </div>
+
+</body>
+</html>
